# server/swagger/swagger.yaml
swagger: '2.0'
info:
  description: Documentado api que será usado no projeto tcc
  version: 1.0.0
  title: Locator API
  contact:
    email: diegoop.pa@hotmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:3001'
basePath: /
...
schemes:
  - http
paths:
...
  /suggestion:
    get:
      tags:
        - Suggestion
      summary: get suggestions
      description: get suggestions
      operationId: getSuggestions
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: limit
          schema:
            type: number
          description: limit for employees
        - in: query
          name: page
          schema:
            type: number
          description: page for employees
        - in: query
          name: field
          schema:
            type: string
          description: get with filter
      responses:
        '200':
          description: Suggestions
          schema:
            type: object
            properties:
                limit:
                  type: number
                page:
                  type: number
                total:
                  type: number
                docs:
                  type: array
                  items:
                   $ref: '#/definitions/Suggestion'        
        '401':
          description: Not Authorzied
          schema:
            $ref: '#/definitions/NotAuthorized'
        '500':
          description: unknown error
          schema:
            $ref: '#/definitions/UnknownServerError'
      security:
      - ApiKey: []

    post:
      tags:
        - Suggestion
      summary: register suggestions
      description: register suggestions
      operationId: register suggestion
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: Suggestion
          schema:
            type: object
            properties:
              author:
                type: string
              title:
                type: string
              onModel:
                type: string
              messages:
                type: array
                items:
                  properties:
                    message:
                      type: string
                    by:
                      type: string
          description: register a new suggestion or question
      responses:
        '200':
          description: response to create suggestion
          schema:
            type: object
            properties:
                message:
                  string: string
        '401':
          description: Not Authorzied
          schema:
            $ref: '#/definitions/NotAuthorized'
        '500':
          description: unknown error
          schema:
            $ref: '#/definitions/UnknownServerError'
      security:
      - ApiKey: []
  
    patch:
      tags:
        - Suggestion
      summary: add message to suggestion
      description: add suggestions
      operationId: addMessage
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: Suggestion
          schema:
            type: object
            properties:
              id:
                type: string
              data:
                type: object
                properties:
                  by:
                    type: string
                  message:
                    type: string
                  onModel:
                    type: string
                    enum: ["User", "Employee"]
          description: add a message to suggestion
      responses:
        '200':
          description: response to create suggestion
          schema:
            type: object
            properties:
                message:
                  string: string
        '401':
          description: Not Authorzied
          schema:
            $ref: '#/definitions/NotAuthorized'
        '500':
          description: unknown error
          schema:
            $ref: '#/definitions/UnknownServerError'
      security:
      - ApiKey: []

    delete:
      tags:
        - Suggestion
      summary: delete suggestion
      description: delete suggestions
      operationId: deleteSuggestion
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: Suggestion
          schema:
            type: object
            properties:
              id:
                type: string
          description: remove suggestion
      responses:
        '200':
          description: response to create suggestion
          schema:
            type: object
            properties:
                message:
                  string: string
        '401':
          description: Not Authorzied
          schema:
            $ref: '#/definitions/NotAuthorized'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/NotFound'
        '500':
          description: unknown error
          schema:
            $ref: '#/definitions/UnknownServerError'
      security:
      - ApiKey: []
  /suggestion/{id}:
    get:
      tags:
        - Suggestion
      summary: get suggestion by id
      description: get suggestion by id
      operationId: getSuggestionById
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          type: string
          description: suggestion id
      responses:
        '200':
          description: suggestion
          schema:
            type: object
            properties:
              schema:
                $ref: '#/definitions/Suggestion'        
        '401':
          description: Not Authorzied
          schema:
            $ref: '#/definitions/NotAuthorized'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/NotFound'
        '500':
          description: unknown error
          schema:
            $ref: '#/definitions/UnknownServerError'
      security:
      - ApiKey: []
...
securityDefinitions:
  ApiKey:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
...
  NotFound:
    type: object
    properties:
      message:
        type: string
        example: entity not found
      status:
        type: integer
        example: 404
  NotAuthorized:
    type: object
    properties:
      message:
        type: string
        example: not authorized
      status:
        type: integer
        example: 401
  UnknownServerError:
    type: object
    properties:
      message:
        type: string
        example: Unexpected error
      status:
        type: integer
        example: 500
...
  Suggestion:
    type: object
    properties:
      _id:
        type: string
      author:
        type: object
        schema:
          $ref: '#/definitions/User'
      title:
        type: string
      createdAt:
        type: string
      messages:
        type: array
        items:
          properties:
            message:
              type: string
            by:
              type: string
            onModel:
              type: string
...